+++ game/view.cpp
@@ -14,26 +14,26 @@ View::View(Stage& inStage,  ScreenManager& inScreen)
     assert(stage.GetHeight() >= screen.GetHeight());
  }
 
-void View::DrawImage(const Vector<int>& at, const class Image& image)
+void View::DrawImage(const Vector<s16>& at, const class Image& image)
 {
     screen.DrawImage(StageToScreen(at), image);
 }
 
-void View::DrawPixel(const Vector<int>& at, u8 color)
+void View::DrawPixel(const Vector<s16>& at, u8 color)
 {
     screen.DrawPixel(StageToScreen(at), color);
 }
 
-void View::DrawRect(const Rect<int>& rect, u8 color)
+void View::DrawRect(const Rect<s16>& rect, u8 color)
 {
-    Vector<int> shifted = rect.origin - screenOffset;
-    Vector<int> wrapped = stage.WrapCoordinate(shifted);
+    Vector<s16> shifted = rect.origin - screenOffset;
+    Vector<s16> wrapped = stage.WrapCoordinate(shifted);
     shifted.x %= stage.GetWidth();
     shifted.y %= stage.GetHeight();
 
-    screen.DrawRect(Rect<int>(shifted, rect.size), color);
-    screen.DrawRect(Rect<int>(wrapped, rect.size), color);
-    screen.DrawRect(Rect<int>(Vector<int>(wrapped.x,shifted.y), rect.size), color);
-    screen.DrawRect(Rect<int>(Vector<int>(shifted.x,wrapped.y), rect.size), color);
+    screen.DrawRect(Rect<s16>(shifted, rect.size), color);
+    screen.DrawRect(Rect<s16>(wrapped, rect.size), color);
+    screen.DrawRect(Rect<s16>(Vector<s16>(wrapped.x,shifted.y), rect.size), color);
+    screen.DrawRect(Rect<s16>(Vector<s16>(shifted.x,wrapped.y), rect.size), color);
 }
